Add a very brief cooldown to the Hookshot.
This adds a 75ms cooldown to the hookshot before it can be fired again after returning. The ability to fire it again so fast was causing a strange bug where it would fire a second time but move at twice the speed. My guess is that the hookshot object was getting duplicated causing two simultaneous update loops which made it move twice the speed. This is just a theory as I really have no idea. I printed out all the values during normal and 2x speed operations and everything was normal. So anyone's guess is as good as mine.