Classic Camera: Fix Spiny Beetle carried object crash.
Trying to carry the object on its back to another field caused the game to crash if the object was "ObjStone": basically rocks, skulls, or pots (though there is no spiny beetle using pots). The problem was it was always trying to despawn the object, and stone's have an update loop that detects if the map is 2D or not. If the object goes null while carried, the map reference goes null, causing a crash.