Classic Camera: Implement camera locking object for transitions.
This implements "camera field" objects on most maps that the classic camera can lock onto during transitions. Because Link can start the second half of a transition out of bounds, the camera did not have a means to lock onto the target field that Link would be entering. Instead, the camera always followed Link and locked onto the center of the current field, even if that field is out of bounds. This would show a black screen with a brief flash of Link walking off screen then snap into the field it was supposed to show.

The implementation of this camera object is to place it near every door in maps that require the object. The closest camera field object to the exit door in the transition is the one the camera temporarily locks onto until the transition is finished. A camera field object can also have a flag set to lock onto the object permanently until returning to the overworld. This is really only useful for single room maps since the camera wouldn't be able to escape past the field of that room.

Finally, this commit removes a hack that was used to cover up the transition glitches since classic camera was first implemented. The hack would draw rectangles over the areas where Link was temporarily visible during transitions. This was an ugly hack that I've wanted to get rid of for so long.