Prevent players from jumping to Marin when stuck on the mountain.
This puts a blocking field in front of Marin that doesn't allow players to jump over the gap in the bridge and forces the usage of the hookshot. There is probably better ways to fix this, but since two iterations of developers didn't bother then I don't see why I should put any more effort into it. This fix is good enough to force the behavior that the original developers intended.