Store original Z position when picking up rooster and apply to hole reset position.
This fixes crosses a hole reset point when flying with the rooster. When picking it up and carrying it, the Z position was saved when setting the hole reset position. With this commit, the original Z position is saved when picking up the rooster, and then that is used when setting the hole reset position.