Classic Camera: Hide Link behind the "circle effect" during first load in.
This hides Link from view until the map has loaded in enough to be able to get the current field. This only seems to be a problem on the "first load" into the world: start new game, loading a save, and resuming from game over. After the first load, we can draw Link over the circle shader again as normal.