Always animate dying enemy and death effects + other stuff.
When knocking an enemy outside the current viewport (especially when classic camera is active), the explosion effect would be visibly frozen until the field is shifted to its location. This commit adds the dying enemy and any spawned effects to the always animate lists.

This also removes the "set Z to 0" from Stalfos enemies which was a workaround that attempted to keep them in the current field. Finally, when adding an object to the always animate list, the list is checked to see if the enemy was already added to it.