Fix cracked tiles to detect pushing and increase move stone push time.
There is a strange issue where pushing against some rocks puts Link in the "pushing" state, while other rocks will put him in the "idle" state. Cracked floors were not looking for "pushing" state, so while he was in that state the floor would never break. This commit makes cracked floors also look for pushing state, and it increases the activation time for stones to 500ms from 450ms so that the timings better match up with the original game.