Implement new Dungeon 4 tile order puzzle system.
The Dungeon 4 tile order puzzle is the one where you have to step on 5 tiles in the correct order. The first set of tiles have a marker to indicate the correct tile. The second puzzle uses the solution from the first but the marker is gone. In the original games, the Dungeon 4 tile puzzle had at least three different solutions. This port only had one solution that was hardcoded via map objects known as "ObjButtonOrder" objects. Each one has the order set as one of their parameters.

The new implementation deletes the ObjButtonOrders off the map, and implements a new "ObjDungeonTilePuzzle" object. The tile puzzle object generates the tile order when entering the dungeon for the first time, and stores that order in the save file so it never changes. It still uses the same underlying system with "ObjButtonOrder" objects, but spawns them in the order matching the order that was generated when entering the dungeon. Both tile puzzle objects share the same solution since you are supposed to use one to solve the other.

This commit also adds a new editor icon specifically for the tile puzzle. So to recap: button order objects were removed, new tile puzzle object was created, tile puzzle object creates a button order when dungeon is entered, there is three possible button orders, button order is stored in save file and never changes, tile puzzle object generates button order objects in the correct order, and both puzzles share the same solution.