Fix strange collision issue when pushing rocks into holes.
The cave on the way to Tal Tal Mountain Range has a rock you can push upward into a hole. Sometimes Link would clip through the stone when pushing the rock. This happened after the last commit that removed `IgnoreHoles = true` which is required to detect hole functions. So this commit instead removes it when the push is nearly finished.