Alternate method to restore music when leaving Dodongo Snake room.
Instead of creating the field with a 1 pixel border smaller than the current room, this instead uses an invisible button that stretches across the outside fields. When the player leaves the field with the snakes, they step on the button which resets the music. This has no chance of failing unlike the pixel border which could be crossed faster than the next update loop could hit.