Implement better "hole absorb" function for pushable stones.
The way to detect a push stone falling down a hole was done via the "pulling" method because I couldn't figure out how to get the "absorb" method to work. This was a horrible way to go about it because pulling happens every update and there is no way to detect when pulling was finished, so whatever it was I wanted to do had to be done just before the last tick. The stone was absorbed way too early this way because I had to disable "no hole absorb" on the final tick before it should be absorbed. This caused some weird bugs like Link teleporting on the other side of the stone, and key-value pairs not getting set because the move never finished.

Also don't respawn move stones for classic camera when created by object spawners. Chances are, the only time they are created by object spawners is when they shouldn't respawn anyway.