Handle Dungeon 7 and Dungeon 8 being completed out of order.
In the possibility that a player completes dungeon 7 and dungeon 8 out of order, this commit tries to set the correct states by checking if the player as the other dungeon's instrument and updating the correct flags. Basically, if dungeon 8 is completed first, do nothing when completing it and do everything when completing 7. And if the order is correct, do everything as it was.