Fix rooster drawn on background layer when borrowing + random.
When borrowing the rooster from the hen house at the end of the game, it was drawn on the background layer. The only way to fix this was to save with the rooster, quit, then reload the game. Now the proper draw layer is set when borrowing. This commit also updates some comments in script.zScript and moves some stuff around in the overworld but should have no functional changes.