Recycle Link's Bow Wow object for world teleportation.
Link has had a reference to Bow Wow when a follower for awhile (_objBowWow) which has gone unused for some time. Since it's there, it's now used to reference Bow Wow when a follower once again for world teleportation. This is not perfect because the chain has limited control over it, but it's better than it was.