Fix Smasher when both it and Link pick up ball at same time.
If Link and Smasher both grabbed the ball at the same time, it would bug out Smasher's animation and bug out the ball which could hurt Link if throwing it in a way where it would collide with him. This fixes both grabbing at the same time by prioritizing Link.