Respawn Hardhat Beetle if knocking it down a hole.
This creates a respawner for Hardhat Beetle that is created when it is knocked down a hole. The beetle will respawn with standard camera when leaving the current field, and it will respawn with classic camera with a field transition. While these two things are very similar, there is a slight difference in that a field transition has several stages while leaving the field is an instantaneous event.