Change the way "pushing" state is determined.
Due to the changes required to make "_isWalking" work for the new "push button" object, it no longer works to determine if Link is walking and colliding because "_isWalking" will always be false here. There's more than one way to determine if he's actually walking, check if he's grounded and his velocity instead.