Add Vacuum to "always animate" list.
Under the right circumstances, a player can get stuck facing in a direction with no way to get out of it. The problem is the vacuum has a "StopRotating" function that needs to be called. But if it's not being rendered, it won't fire. So by adding it to always animate, it will always be called.