Rearrange Pairodd "OnHit" function so components get disabled.
When dashing at a Pairodd with Pegasus Boots and the sword equipped, the Pairodd will die instantly but the dying body will bounce the player back because the shield "bumps" it. This is due to the order of operations on the OnHit function where when it was dying, disabling the components never happened.