Allow configuring ObjPushButton "destroy" value.
This makes it possible to specify the value the new "ObjPushButton" uses to destroy itself, and implements the object in Level 5 dungeon for the doorways to Master Stalfos. Because Master Stalfos uses an incrementing value, it was not possible to apply push buttons to the doorway of subsequent fights.

So in summary: this commit updates the push button so the destroy value can be specified, it adds the push buttons to Master Stalfos doorways in dungeon 5, and it updates all previous maps so the destroy value is properly implemented.