Jump fix hack messes up Lvl2 sword grab.
The amount of pain and suffering this hack has caused me is endless. I wish I could think of a better way to not "slide" when landing but this is the best I could do. Unfortunately it screws up so many animations that require workarounds to the workaround. I keep praying there is an end to it eventually.