Don't force successful block on Sea Urchin.
I don't know why this was true, but going back to v1.0.0 it was true there too. But, until v1.3.3 Sea Urchins could damage Link from the side when blocking. This was the version that shield changes were made more accurate, so its possible a forced block caused it to always succeed when before it wouldn't have.