Urchins deal damage from the side when blocking.
This was difficult to pull off because they don't have a damage component. For the longest time, it has relied on the push component to deal damage. But walking into an urchin from the side does not register a push. This commit adds a complementary function using the collision to detect "touching from the side". It's not 100% perfect, but it's closer to accurate behavior than not.