Fix owl freeze trigger, rooster carry, and inventory blocking.
I used to just check if all the instruments were obtained to know if the owl should stop freezing the game world since it can soft lock the player at the egg or during the ending. But instead it's possible to just use the key associated with the owl in those positions. Also use a backup method of disabling the inventory for times when something else set's it to false outside of the owl object. Finally, when the owl enters the map, drop the rooster.