More ObjLink, ObjLink2d, and ObjLinkHelper "cleanups".
All of these are ultimately part of just "ObjLink". While I have mostly finished the cleanups of ObjLink and ObjLinkHelper, and moved a lot of "one off" logic to the latter, ObjLink2d still remains kind of a mess. Something deleted my white space in ObjLink somewhere along the line, so this restores it in places where it looks nicer and makes things easier to read.