Classic Camera: Only respawn objects on the previous field.
Objects that respawn (bushes, grass, stones, ice blocks, crystals) are handled by "respawner" objects that detect a field change and then respawn the object they are responsible for respawning. The problem with this is that it's possible to destroy objects outside of Link's current field, so it's weird to see objects respawn on the field he is crossing into. This commit makes it so objects only respawn when they are on the previous field, or rather the field that he just came from, and not the field he is crossing into.