Classic Camera: Fix resetting objects at the bottom of the field.
When an object exists on or is pushed to the bottom of the current field, their "EntityPosition" can technically fall outside of the current field. This prevents them from properly being reset because they are not technically on the field. To fix this, extend the "PreviousField" by 4 pixels downward to compensate of objects that are still on the field but their position states otherwise.