Properly "always animate" the boomerang and followers.
There were still two remaining holdouts that were always animated but not through the "always animate" list. They were manually being added to the groups of objects that get animated instead of being added to the lists which have several protections to make sure no issues occur. This is a behavior change, so there could potentially be new issues. It does however fix these objects from sticking around as "ghost objects" when quitting the game while they are active on the map.