Push tweaks in "SystemBody.cs" and remove Urchin hack.
This does something with pushing, I'm not exactly sure what but it fixes at least two known issues with pushing not being detected when running at a 45 degree angle:

- Push not being detected when pushing Sea Urchins diagonally. The push would fail to register and no damage is taken.
- Push not being detected when triggering the raccoon "lost" message. This uses a weird object called a "push dialog" which detects a push for one frame.

This commit also removes the hack I created for Urchins that tries to use collision to detect when to deal damage. This method was heavily flawed, and is no longer needed with the fix in SystemBody.