Rework Octorok shot cooldown.
The way the cooldown was working was completely bunk. It was almost as if they never even fired anymore. This was especially noticeable with classic camera. If they aren't on screen, their cooldown was not reducing. And it only reduced slightly when they walked. Whatever the case, it was not working.