Isolate deadzone to only affecting analog stick.
It was being applied to D-Pad input as well. Deadzone was also used everywhere else to determine if Input was being entered, which IMO is a wrong way to do it. It should only affect the initial input of the analog stick, not be used as a comparison tool later.