Update overworld map with mountain photo fixes.
This is an addendum to ca6a22d because I forgot to upload the overworld changes. I went to test the photo and it didn't work, so I looked and the changes were gone. I must have forgot to copy them from the build to the source code folder and then overwrote them. Derp.