Fix file select animations that have sword 2 only.
Files that have the Level 2 sword but not the Mirror Shield only show the sword in hand. There was an animation bug where the sword would bounce back and forth between frames. This was caused by an offset added to the second frame, so that offset has been eliminated and instead compensated for by making the frame 1 pixel larger.