Dungeon teleporter doesn't teleport when spawning on player.
When spawning it, start it in a "locked" state. Logically, this should disable it until the player touches it twice, but it only seems to lock it when it spawns on the player. And if it doesn't spawn on the player, the first touch works? Don't know how or why but since it works I don't care either.