Add new Modifiiers option "Sword Item Interaction".
The ability to whack bombs with the sword has been moved to this option, as well as being able to collect fairies. Also a new feature has been added where the boomerang will bounce off the sword when this option is enabled. Perhaps more features will be added to it in the future.