Fix Raven fly triggering in ending sequence.
The short of it is, Ravens are always animated. When map is loaded, Link is put at position 0,0 just invisible. This triggers the Raven which causes it to fly, which plays the flapping sound. So to prevent this, when the map is loaded for the ending, just remove all the Ravens.