Fix up dungeon 5 boss Slime Eel inconsistencies.
This commit adds a delay between when it smashes out the four walls and when performing the first attack. In the original games, there was about a 7 second delay while the player was forced to dodge the tail until it starts to attack. That functionality has been restored here. I just cloned the "hidden" state and created a "start" state which does the same thing, except it takes longer and is only called once.

Since the explosion sound fix it would also play the death throes sound effect twice: once before the dialog, and then again after. So I implemented a variable to separate the death sound from the death explosion sound. On Slime Eel, the death sound is played before the dialog, so it must be done on the object itself instead of in AIDamageState like most bosses.