Fix final boss Zol form from being always vulnerable.
There was never any checks in place to detect when it should be able to be hit by powder, so it was possible to just spam it at any point and basically insta-kill it. This commit makes it so it's only vulnerable when above ground and jumping around, and will not take a second hit after the first until it burrows and comes back up.