Redesign the shopkeeper's revenge sequence.
This makes the shopkeepers revenge behave much closer to the original game. When trying to steal, instead of a 50/50 chance of turning around, the turn-around delay is a random value between 0ms and 2500ms. During revenge, health is reduced over time instead of all at once and the items are hidden from view. This also fixes the bug where Link would not die if the Damage Multiplier is set below 0.50x.