Animator stores AnimationID and fix 2D quick land + jump.
The animator doesn't store the string of the animation sent to play so store it when running a "Play" command. This is used to fix 2D when landing and jumping immediately after. The jump animation is still playing, so it fails to play it again, so check if the current animation is still jumping and fail until it is.