Update fonts: Cyrillic. Fix invalid characters.
This replaces the Cyrillic (Russian) characters with new fonts for both the original game and the variable width font. The original font is made to appear like the original font for the GameBoy while the VWF is a close match for the VWF used in the Redux mod. This also fixes crashes on certain screen in French and Spanish languages due to the use of invalid characters.